# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Freight Management Platform API
  description: |
    Comprehensive API for Ethiopian freight logistics platform.

    ## Authentication
    Most endpoints require authentication via session cookie or Bearer token.
    - **Web clients**: Use session cookie with CSRF token in X-CSRF-Token header
    - **Mobile clients**: Use Bearer token in Authorization header

    ## Rate Limiting
    All endpoints are rate-limited. When exceeded, returns 429 with Retry-After header.

    ## Error Handling
    - 400: Validation error (Zod schema)
    - 401: Unauthorized
    - 403: Forbidden (insufficient permissions)
    - 404: Resource not found
    - 409: Conflict (state machine violation)
    - 429: Rate limit exceeded
    - 500: Internal server error
  version: 1.0.0
  contact:
    name: Freight Platform Support
    email: support@freightplatform.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Development
  - url: https://api.freightplatform.com/api
    description: Production
...
paths:
...
  /trucks/{id}/nearby-loads:
    get:
      tags: [trucks, matching]
      summary: Find nearby loads for truck
      operationId: findNearbyLoads
      parameters:
        - $ref: "#/components/parameters/TruckId"
        - name: radiusKm
          in: query
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: Nearby matching loads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoadMatch"
...
  /trips/{tripId}/history:
    get:
      tags: [trips, gps]
      summary: Get trip route history
      operationId: getTripHistory
      parameters:
        - $ref: "#/components/parameters/TripId"
      responses:
        "200":
          description: Route history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoutePoint"

  # ============================================================================
  # WALLET
  # ============================================================================
...
  /gps/eta:
    get:
      tags: [gps]
      summary: Get ETA for trip
      operationId: getETA
      parameters:
        - name: tripId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ETA information
          content:
            application/json:
              schema:
                type: object
                properties:
                  etaMinutes:
                    type: integer
                  distanceRemainingKm:
                    type: number
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  parameters:
...
    TruckId:
      name: id
      in: path
      required: true
      schema:
        type: string
    TripId:
      name: tripId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    TruckType:
      type: string
      enum:
        [
          FLATBED,
          REFRIGERATED,
          TANKER,
          CONTAINER,
          DRY_VAN,
          LOWBOY,
          DUMP_TRUCK,
          BOX_TRUCK,
        ]
    LoadStatus:
      type: string
      enum:
        [
          DRAFT,
          POSTED,
          UNPOSTED,
          SEARCHING,
          OFFERED,
          ASSIGNED,
          PICKUP_PENDING,
          IN_TRANSIT,
          DELIVERED,
          COMPLETED,
          CANCELLED,
          EXPIRED,
          EXCEPTION,
        ]
...
    OrganizationSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        isVerified:
          type: boolean

    # --------------------------------------------------------------------------
    # LOAD
    # --------------------------------------------------------------------------
    Load:
      type: object
      properties:
        id:
          type: string
        status:
          $ref: "#/components/schemas/LoadStatus"
        pickupCity:
          type: string
        deliveryCity:
          type: string
        pickupDate:
          type: string
          format: date-time
        deliveryDate:
          type: string
          format: date-time
        truckType:
          $ref: "#/components/schemas/TruckType"
        weight:
          type: number
        cargoDescription:
          type: string
        isAnonymous:
          type: boolean
        ageMinutes:
          type: integer
        shipper:
          $ref: "#/components/schemas/OrganizationSummary"
...
    LoadMatch:
      type: object
      properties:
        load:
          $ref: "#/components/schemas/Load"
        matchScore:
          type: integer
          minimum: 0
          maximum: 100
        dhOriginKm:
          type: number
        matchReasons:
          type: array
          items:
            type: string

    # --------------------------------------------------------------------------
    # TRUCK
    # --------------------------------------------------------------------------
...
    RoutePoint:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
        speed:
          type: number
        heading:
          type: number
        timestamp:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
  - cookieAuth: []